✨ feat(api.dart): add CharacterImageType enum and getImageUrl function The CharacterImageType enum is added to provide a list of available image types for a character. The getImageUrl function is added to generate the URL for a specific image type of a character.
🐛 fix(character_detail_cubit.dart): rename getCharacterDetailCard to getCharacterDetail
The function name is changed to better reflect its purpose.

🐛 fix(character_datasource.dart): add getCharacterCard function
The getCharacterCard function is added to retrieve the character card image from the API.

🐛 fix(character_response_model.dart): change birthday type to String
The birthday field is changed to a String type to match the API response.

✨ feat(character_details_page.dart): add character detail page
The character detail page is added to display the character card image.

✨ feat(character_home_page.dart): add crossAxisSpacing to GridView
The crossAxisSpacing property is added to