🎨 style(api.dart): add gachaSplash to CharacterImageType enum 🚀 perf(api.dart): add getGachaSplashUrl method to GenshinConst class The `CharacterImageType` enum has been updated to include a new value `gachaSplash`. The `GenshinConst` class has been updated to include a new method `getGachaSplashUrl` which returns the URL for the `gachaSplash` image type.
🚀 perf(character_details_page.dart): add navigation to character more information page
The `_openDetail` method has been added to `_CharacterDetail` class to navigate to the `GenshinCharacterMoreInformationPage` when the user taps on the character. This improves the user experience by allowing them to view more information about the character.

🎨 style(character_home_page.dart): update childAspectRatio of _CharacterCard
The `childAspectRatio` of `_CharacterCard` has been updated to

🎨 style(character_more_information_page.dart): format code
🚀 chore(character_more_information_page.dart): add more information page for Genshin characters
🚀 chore(character_overlay.dart): add hero tags for each hero widget
🚀 chore(thunderEnvironment.json): change character name to Klee
The code has been formatted to improve readability. A new page has been added to display more information about Genshin characters. The page includes a hero image, a back button, and a gradient background. Each hero widget now has a unique hero tag to prevent conflicts. The character name in the test environment has been changed to Klee.